Channel (elixtor v0.2.0-beta)

The Channel module handles the communication between nodes in a network.

It provides functions to send and receive cells over a secure SSL socket, and other functions related to establishing and using a channel.

As of now, link protocols other than version 3 are not explicitly supported.

You can read more about Channels in the Tor spec: https://spec.torproject.org/tor-spec/channels.html

Types

t()

@type t() :: %Channel{circ_id_len: 16 | 32, tls_socket: :ssl.socket()}

A channel is a TLS connection to a Tor relay, the width of the cird_id is negotiated during the handshake.

Functions

check_tls_hash(channel, signing_v_tls_cert)

@spec check_tls_hash(t(), Channel.Parsing.Certs.t()) :: :ok | {:error, any()}

Validates the TLS certificate of the given channel against the given signing certificate.

This function DOES NOT validate the certificate chain, it only checks if the subject matches.

convert_and_send(ch, cell, cell_mod)

@spec convert_and_send(t(), struct(), Channel.CellBehaviour) ::
  :ok | {:error, :ssl.reason()}

Converts a cell of a specific type into a binary cell and sends it over the given channel.

The cell is converted to a binary cell by the given cell module.

For example, to send a VERSIONS cell:

cell = Channel.Cells.Versions.new([1, 2, 3])
Channel.convert_and_send(channel, cell, Channel.Cells.Versions)

new(tls_socket)

@spec new(:ssl.socket()) :: t()

Creates a new channel with the given TLS socket.

recv_and_convert(ch, cell_mod)

@spec recv_and_convert(t(), Channel.CellBehaviour) ::
  {:ok, struct()} | {:error, :ssl.reason()}

Receives a cell and converts it to a struct using the given cell module.

For example, to receive a VERSIONS cell:

{:ok, versions_cell} = Channel.recv_and_convert(channel, Channel.Cells.Versions)

recv_cell(channel)

@spec recv_cell(t()) :: {:ok, Channel.BinaryCell.t()} | {:error, :ssl.reason()}

Receives the next cell from the given channel in the order they were received.

send_cell(channel, cell)

@spec send_cell(t(), Channel.BinaryCell.t()) :: :ok | {:error, :ssl.reason()}

Sends the given cell over the given channel.