Channel.CellBehaviour behaviour (elixtor v0.1.0-beta)

Defines behaviour for converting between binary cells and specific cell types. Cells are differentiated by their command code.

Tor provides a list of cell types with hyperlinks to the relevant sections of the Tor specification in section 2.3.2 "Cells (messages and channels)".

Callbacks

from_binary_cell(t)

@callback from_binary_cell(Channel.BinaryCell.t()) :: {:ok, struct()} | {:error, any()}

from_keywords(keyword)

@callback from_keywords(keyword()) :: {:ok, struct()} | {:error, any()}

Construct a cell of the given type given a keyword list of options.

to_binary_cell(struct)

@callback to_binary_cell(struct()) :: {:ok, Channel.BinaryCell.t()} | {:error, any()}

Convert a cell of the given type to a binary cell.